+++ extensions/foo_jl_biography_mac/src/Core/BiographyData.mm
@@ -140,6 +140,32 @@
     return self;
 }
 
+- (instancetype)initWithData:(BiographyData *)data {
+    self = [self init];
+    if (self) {
+        _artistName = [data.artistName copy];
+        _musicBrainzId = [data.musicBrainzId copy];
+        _biography = [data.biography copy];
+        _biographySummary = [data.biographySummary copy];
+        _biographySource = data.biographySource;
+        _language = [data.language copy];
+        _artistImage = data.artistImage;
+        _artistImageURL = [data.artistImageURL copy];
+        _imageSource = data.imageSource;
+        _imageType = data.imageType;
+        _tags = [data.tags copy];
+        _similarArtists = [data.similarArtists copy];
+        _genre = [data.genre copy];
+        _country = [data.country copy];
+        _listeners = data.listeners;
+        _playcount = data.playcount;
+        _fetchedAt = data.fetchedAt;
+        _isFromCache = data.isFromCache;
+        _isStale = data.isStale;
+    }
+    return self;
+}
+
 - (BiographyData *)build {
     NSAssert(self.artistName.length > 0, @"Artist name is required");
     return [[BiographyData alloc] initWithBuilder:self];